Added Shader, VertexArray and Buffer

+++ VoxelEngine/src/Platform/Windows/WindowsWindow.h
@@ -14,8 +14,8 @@ namespace VoxelEngine
 
 		void OnUpdate() override;
 
-		inline unsigned int GetWidth() const override { return m_Data.Width; }
-		inline unsigned int GetHeight() const override { return m_Data.Height; }
+		inline uint16_t GetWidth() const override { return m_Data.Width; }
+		inline uint16_t GetHeight() const override { return m_Data.Height; }
 
 		inline void SetEventCallback(const EventCallbackFn& callback) override { m_Data.EventCallback = callback; }
 
@@ -34,7 +34,7 @@ namespace VoxelEngine
 		struct WindowData
 		{
 			std::string Title;
-			unsigned int Width, Height;
+			uint16_t Width, Height;
 			bool VSync;
 
 			EventCallbackFn EventCallback;
