+++ VoxelEngine/src/VoxelEngine/Window.h
@@ -11,12 +11,12 @@ namespace VoxelEngine
 	struct WindowProps
 	{
 		std::string Title;
-		unsigned int Width;
-		unsigned int Height;
+		uint16_t Width;
+		uint16_t Height;
 
 		WindowProps(const std::string& title = "Voxel Engine",
-			unsigned int width = 1280,
-			unsigned int height = 720)
+			uint16_t width = 1280,
+			uint16_t height = 720)
 			: Title(title), Width(width), Height(height)
 		{
 		}
@@ -31,8 +31,8 @@ namespace VoxelEngine
 
 		virtual void OnUpdate() = 0;
 
-		virtual unsigned int GetWidth() const = 0;
-		virtual unsigned int GetHeight() const = 0;
+		virtual uint16_t GetWidth() const = 0;
+		virtual uint16_t GetHeight() const = 0;
 
 		virtual void SetEventCallback(const EventCallbackFn& callback) = 0;
 		virtual void SetVSync(bool enabled) = 0;
